Добавляет геренацию жанров через map. Небольшие улучшения читаемости кода.
