+++ src/index.js
@@ -1,43 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/app/app';
-
-const CURRENT_MOVIE = { // МОК-данные с примером одного фильма для рендера главной страницу
-  "name": `The Grand Budapest Hotel`,
-  "poster_image": `img/the-grand-budapest-hotel-poster.jpg`,
-  "preview_image": `img/the-grand-budapest-hotel.jpg`,
-  "background_image": `img/bg-the-grand-budapest-hotel.jpg`,
-  "genre": `Comedy`,
-  "released": 2014,
-};
-
-const MOVIES_LIST = { // МОК-данные с примерами фильмов
-  "Fantastic Beasts: The Crimes of Grindelwald": `fantastic-beasts-the-crimes-of-grindelwald.jpg`,
-  "Bohemian Rhapsody": `bohemian-rhapsody.jpg`,
-  "Macbeth": `macbeth.jpg`,
-  "Aviator": `aviator.jpg`,
-  "We need to talk about Kevin": `we-need-to-talk-about-kevin.jpg`,
-  "What We Do in the Shadows": `what-we-do-in-the-shadows.jpg`,
-  "Revenant": `revenant.jpg`,
-  "Johnny English": `johnny-english.jpg`,
-  "Shutter Island": `shutter-island.jpg`,
-  "Pulp Fiction": `pulp-fiction.jpg`,
-  "No Country for Old Men": `no-country-for-old-men.jpg`,
-  "Snatch": `snatch.jpg`,
-  "Moonrise Kingdom": `moonrise-kingdom.jpg`,
-  "Seven Years in Tibet": `seven-years-in-tibet.jpg`,
-  "Midnight Special": `midnight-special.jpg`,
-  "War of the Worlds": `war-of-the-worlds.jpg`,
-  "Dardjeeling Limited": `dardjeeling-limited.jpg`,
-  "Orlando": `orlando.jpg`,
-  "Mindhunter": `mindhunter.jpg`,
-  "The Grand Budapest Hotel": `bg-the-grand-budapest-hotel.jpg`,
-};
+import {CURRENT_MOVIE, MOVIES_LIST, GENRES} from './mock/movies.js';
 
 ReactDOM.render(
     <App
       currentMovie = {CURRENT_MOVIE}
       moviesList = {MOVIES_LIST}
+      genres = {GENRES}
     />,
     document.querySelector(`#root`)
 );
